PullStackUpdate_User's Videos Bugfix

+++ src/controllers/videoController.js
@@ -25,19 +25,28 @@ export const watch = async (req, res) => {
     };
 export const getEdit = async(req, res) => {
     const { id } = req.params;
+    const {user: {_id}} = req.session;
     const video = await Video.findById(id);
     if(!video) {
         return res.status(404).render("404" ,{ pageTitle: "Video not found."});
-        }
-    return res.render("edit", {pageTitle:`Edit: ${video.title}`, video });
+    }
+    console.log(typeof video.owner, typeof _id);
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+    return res.render("edit", {pageTitle: `Edit: ${video.title}`, video });
 };
 export const postEdit = async(req, res) => {
     const { id } = req.params;
+    const {user: {_id}} = req.session;
     const {title, description, hashtags} = req.body;
     const video = await Video.exists({ _id: id});
     if(!video) {
         return res.status(404).render("404" ,{ pageTitle: "Video not found."});
-        }
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
         await Video.findOneAndUpdate(id, {
             title, description, hashtags:Video.formatHashtags(hashtags)
         });
@@ -75,6 +84,14 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async(req, res) => {
     const { id } = req.params;
+    const {user: {_id}} = req.session;
+    const video = await Video.findById(id);
+    if (!video){
+        return res.status(404).render("404" ,{ pageTitle: "Video not found."});
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     // delete video
     return res.redirect("/");
